+++ src/App.tsx
@@ -1,64 +1,18 @@
-import { useState, useRef, useCallback } from "react";
-import styled, { createGlobalStyle } from "styled-components";
-import ChatList from "./components/ChatList";
-import InputForm from "./components/InputForm";
-import UserList from "./components/UserList";
-import { Chat, User } from "./interface";
+import { Route, Routes } from "react-router-dom";
+import FriendListPage from "./pages/FriendListPage";
+import ChatroomListPage from "./pages/ChatroomListPage";
+import SettingPage from "./pages/SettingPage";
+import Chatroom from "./pages/Chatroom";
 
-import userData from "./data/userData.json";
-import chatData from "./data/chatData.json";
-
-function App() {
-  const [curUser, setCurUser] = useState(1);
-  const [users, setUsers] = useState<User[]>(userData.users);
-  const [chats, setChats] = useState<Chat[]>(chatData.chats);
-
-  const nextChatId = useRef(chatData.chats.length + 1);
-  const onConcat = useCallback(
-    (text: string) => {
-      const chat = {
-        id: nextChatId.current,
-        senderId: curUser,
-        text,
-        date: String(new Date()),
-      };
-      setChats(chats.concat(chat));
-      nextChatId.current++;
-    },
-    [chats, curUser]
-  );
-
-  const changeUser = (id: number) => {
-    setCurUser(id);
-  };
-
+const App = () => {
   return (
-    <Wrapper>
-      <GlobalStyle />
-      <UserList curUser={curUser} users={users} changeUser={changeUser} />
-      <ChatList curUser={curUser} users={users} chats={chats} />
-      <InputForm onConcat={onConcat} />
-    </Wrapper>
+    <Routes>
+      <Route path="/" element={<FriendListPage />} />
+      <Route path="chatrooms" element={<ChatroomListPage />} />
+      <Route path="setting" element={<SettingPage />} />
+      <Route path="chatroom" element={<Chatroom />} />
+    </Routes>
   );
-}
-
-const GlobalStyle = createGlobalStyle`
-html{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: whitesmoke;
-  font-size: 0.8rem;
-}
-`;
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 350px;
-  height: 95vh;
-  border: 1px solid lightgray;
-  border-radius: 20px;
-`;
+};
 
 export default App;
