+++ src/pages/ChatroomListPage.tsx
@@ -1,0 +1,9 @@
+const ChatroomListPage = () => {
+  return (
+    <div>
+      <h1>채팅방목록</h1>
+    </div>
+  );
+};
+
+export default ChatroomListPage;
