+++ src/pages/FriendListPage.tsx
@@ -1,0 +1,9 @@
+const FriendListPage = () => {
+  return (
+    <div>
+      <h1>친구목록</h1>
+    </div>
+  );
+};
+
+export default FriendListPage;
