+++ src/pages/SettingPage.tsx
@@ -1,0 +1,9 @@
+const SettingPage = () => {
+  return (
+    <div>
+      <h1>설정 더보기</h1>
+    </div>
+  );
+};
+
+export default SettingPage;
